+++ RELEASE-NOTES.md
@@ -18,6 +18,7 @@ as release notes for end user on a Recalbox upgrade.
 - New Random game picker in demo mode (less cycling)
 - Add BSNES core (performance build) (PC and RPi4/400)
 - Add BSNES HD for wide screen Super Nintendo games (PC and RPi4/400)
+- Add Pi400 power button support (Fn + F10 to turn off)
 
 ### Improvements